requirements.txt
brian2
numpy
matplotlib
pyserial